refactor: move out common methods
